// client/types/product.ts
export interface Product {
    id: string;
    name: string;
    description: string;
    price: number;
    power?: string;
    quantity?: number;
    brand?: string;
    weight?: number;
    height?: number;
    width?: number;
    length?: number;
    model_code?: string;
    colour?: string;
    img_url?: string;
}